ploughing through the test rewrite  🐿 v2.5.16
